refactor(example): extract choose image button
